moved config to src
