+++ packages/amplify_authenticator/lib/src/blocs/auth/auth_bloc.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_authenticator/src/blocs/auth/auth_data.dart';
 import 'package:amplify_authenticator/src/services/amplify_auth_service.dart';
-import 'package:amplify_flutter/config/amplify_config.dart';
+import 'package:amplify_flutter/src/config/amplify_config.dart';
 
 //Models
 import 'package:amplify_authenticator/src/models/authenticator_exceptions.dart';
