+++ packages/amplify_authenticator/lib/src/services/amplify_auth_service.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify.dart';
-import 'package:amplify_flutter/config/amplify_config.dart';
+import 'package:amplify_flutter/src/config/amplify_config.dart';
 
 abstract class AuthService {
   Future<SignInResult> signIn(String username, String password);
