+++ packages/amplify_flutter/lib/src/config/analytics/pinpoint_analytics_config.dart
