+++ packages/amplify_flutter/lib/src/config/analytics/pinpoint_targeting_config.dart
