+++ packages/amplify_flutter/lib/src/config/auth/appsync_auth_config.dart
