+++ packages/amplify_flutter/lib/src/config/auth/credential_provider_config.dart
