+++ packages/amplify_flutter/lib/src/config/auth/login_mechanisms.dart
