+++ packages/amplify_flutter/lib/src/config/auth/oauth_scopes.dart
