+++ packages/amplify_flutter/lib/src/config/auth/password_policy_characters.dart
