+++ packages/amplify_flutter/lib/src/config/auth/pinpoint_analytics_auth_config.dart
