+++ packages/amplify_flutter/lib/src/config/auth/pinpoint_targeting_auth_config.dart
