+++ packages/amplify_flutter/lib/src/config/auth/s3_transfer_utility_auth_config.dart
