+++ packages/amplify_flutter/lib/utils/create_config_object.dart
@@ -16,7 +16,7 @@
 import 'dart:convert';
 
 import 'package:amplify_flutter/categories/amplify_categories.dart';
-import 'package:amplify_flutter/config/amplify_config.dart';
+import 'package:amplify_flutter/src/config/amplify_config.dart';
 
 /// Returns an AmplifyConfig object populated only with plugin data
 /// where plugins have been successfully added.
